Ex00/fix negative exchange rate (#9)
* ex00/Fix negative exchange rate

* Update ex01 RPN

Change the return type of ParseNumber_ to int

+++ ex01/RPN.hpp
@@ -8,7 +8,7 @@
 class RPN {
 private:
   typedef std::pair<double, double> NumbersPair;
-  double ParseNumber_(const std::string &token) const;
+  int ParseNumber_(const std::string &token) const;
   NumbersPair GetNumbers_(std::stack<double> *s) const;
 
 public:
